refactor(common): extract internal utilities to their packages
- Delete unused run-command.ts (create-magek has its own implementation)
- Move promises.ts to @magek/core/src/utils/ (internal utility)
- Move internal-info.ts to @magek/server (server-only utility)
- Update CLAUDE.md with GitHub CLI tip

Analysis revealed most utilities in @magek/common are genuinely shared
(used by adapters), so they remain in common. Only truly internal
utilities were moved.

Closes #713
