+++ packages/core/src/subscribers-notifier.ts
@@ -7,9 +7,9 @@ import {
   SubscriptionEnvelope,
   GraphQLData,
   TraceActionTypes,
-  Promises,
   getLogger,
 } from '@magek/common'
+import { Promises } from './utils/promises'
 import { GraphQLSchema, DocumentNode } from 'graphql'
 import * as graphql from 'graphql'
 import { GraphQLGenerator } from './services/graphql/graphql-generator'
